fix(test): minor fixes on database tests
- correct the behavior of some functions
- update the ci accordingly
- refactor .env.example

+++ src/tests/test_db.py
@@ -38,7 +38,8 @@ unique_id = ""
 def configure_database():
     connection = connect(ReDB_HOST, ReDB_PORT, user=ReDB_USER, password=ReDB_PASS)
     create_table(database, empty_table, connection)
-    delete_all(database, table, connection)
+    if len(get_all(database, table, connection)) > 0:
+        delete_all(database, table, connection)
     data = [
         {
             "name": "The Dark Knight Rises",
@@ -86,7 +87,7 @@ def configure_database():
     result = insert(database, table, data, connection)
 
     global unique_id
-    unique_id = result["generated_keys"][0]
+    unique_id = result.get("generated_keys")[0]
 
 
 @pytest.fixture
@@ -200,7 +201,7 @@ def test_unsuccessful_get_all(rethink_connect):
     result = get_all(database, empty_table, rethink_connect)
     assert result is not None
     assert len(result) == 0
-    assert result == {}
+    assert result == []
 
 
 def test_successful_get(rethink_connect):
